got map working
