+++ src/js/keys.js
@@ -1,0 +1,3 @@
+var GMAPS_KEY = "AIzaSyDgHG7eIe9ZL_IAjZuy0ZkqDe6nBdiG03Q";
+
+export {GMAPS_KEY};
