+++ src/js/main.js
@@ -1,15 +1,10 @@
 import {getNews} from "./getApi.js";
 import $ from "jquery";
+import {GMAPS_KEY} from "./keys.js"
+import {addMap, populateNews} from "./generate.js";
 
 var data = getNews();
 
 data.then(populateNews);
 
-function populateNews(results){
-	var newsHtml = 
-	`
-	<span class="news-title">${results.title}</span><span class="news-date">${results.date_published}</span>
-	<p class="news-post">${results.post}</p>
-	`
-	$(".news").html(newsHtml);
-};
+addMap(GMAPS_KEY);
